% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roulette.R
\name{pkg_roulette}
\alias{pkg_roulette}
\title{CRAN package name roulette}
\usage{
pkg_roulette(time = 4, n = 100, cols = sample(colors(), n, replace = TRUE))
}
\arguments{
\item{time}{How long in seconds the roulette runs. It will increasingly speed up.}

\item{n}{number of packages in the roulette, packages are selected with \code{\link[=available.packages]{available.packages()}}}

\item{cols}{colors. The default is to pick \code{n} colors
from the built-in \code{\link[grDevices:colors]{grDevices::colors()}}.}
}
\value{
the last package, invisibly
}
\description{
Runs a roulette of available packages on CRAN
}
\examples{
\dontrun{
  pkg_roulette(time = 3, n = 100)
}

}
