% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{verse}
\alias{verse}
\title{Write poems about R packages}
\usage{
verse(pkg = "ggplot2")
}
\arguments{
\item{pkg}{A package to write a poem about}
}
\description{
Write poems about R packages
}
\examples{
\dontrun{
  chattr::chattr_use("gpt35")
  tada::verse("knitr")
}

}
